+++ packages/kokkos/core/src/impl/Kokkos_MemoryPool.cpp
@@ -41,135 +41,4 @@
 //@HEADER
 */
 
-#ifndef KOKKOS_MEMORYPOOL_CPP
-#define KOKKOS_MEMORYPOOL_CPP
-
-//----------------------------------------------------------------------------
-
-#if defined( KOKKOS_ACTIVE_EXECUTION_MEMORY_SPACE_CUDA )
-
-/* This '.cpp' is being included by the header file
- * to inline these functions for Cuda.
- *
- *  Prefer to implement these functions in a separate
- *  compilation unit.  However, the 'nvcc' linker
- *  has an internal error when attempting separate compilation
- *  (--relocatable-device-code=true)
- *  of Kokkos unit tests.
- */
-
-#define KOKKOS_MEMPOOLLIST_INLINE inline
-
-#else
-
-/*  This '.cpp' file is being separately compiled for the Host */
-
-#include <Kokkos_MemoryPool.hpp>
-#include <Kokkos_Atomic.hpp>
-
-#define KOKKOS_MEMPOOLLIST_INLINE /* */
-
-#endif
-
-//----------------------------------------------------------------------------
-
-#define KOKKOS_MEMPOOLLIST_LOCK  reinterpret_cast<void*>( ~uintptr_t(0) )
-
-namespace Kokkos {
-namespace Experimental {
-namespace Impl {
-
-KOKKOS_FUNCTION
-KOKKOS_MEMPOOLLIST_INLINE
-void * MemPoolList::allocate( size_t arg_size ) const
-{
-  // Requires requested size less than or equal chunk size
-  if ( m_chunk_size < arg_size ) {
-    Kokkos::abort("MemoryPool::allocate ARGUMENT ERROR");
-  }
-
-  void * p = 0 ;
-
-  bool pending = true ;
-
-  while ( pending ) {
-
-    // If head is null then nothing left to allocate.
-    pending = 0 != ( p = *m_head_list );
-
-    if ( pending && ( p != KOKKOS_MEMPOOLLIST_LOCK ) ) {
-      // In the initial look at the head, the freelist wasn't empty or
-      // locked. Attempt to lock the head of list.  If the list was changed
-      // (including being locked) between the initial look and now, head will
-      // be different.  This means the removal can't proceed
-      // and has to be tried again.
-
-      pending = p != atomic_compare_exchange( m_head_list, p, KOKKOS_MEMPOOLLIST_LOCK );
-    }
-  }
-
-  if ( p ) {
-    // The lock succeeded.  Get a local copy of the next entry
-    void * const head_next = *reinterpret_cast<void * volatile *>( p );
-
-    // Replace the lock with the next entry on the list.
-    if ( KOKKOS_MEMPOOLLIST_LOCK != atomic_compare_exchange( m_head_list, KOKKOS_MEMPOOLLIST_LOCK, head_next ) ) {
-      Kokkos::abort("MemoryPool::allocate UNLOCK ERROR");
-    }
-  }
-
-  return p ;
-}
-
-
-KOKKOS_FUNCTION
-KOKKOS_MEMPOOLLIST_INLINE
-void MemPoolList::deallocate( void * arg_alloc , size_t arg_size ) const
-{
-  {
-    // Requires original pointer, test that it is one of the chunks
-
-    const size_t dist = ((char *)arg_alloc) - ((char*)m_head_list);
-
-    if ( ( ( dist % m_chunk_size ) != 0 ) ||
-         ( m_chunk_count < ( dist / m_chunk_size ) ) ||
-         ( m_chunk_size < arg_size ) ) {
-      Kokkos::abort("MemoryPool::deallocate ARGUMENT ERROR");
-    }
-  }
-
-  bool pending = true ;
-
-  while ( pending ) {
-
-    void * const head = *m_head_list ;
-
-    if ( head != KOKKOS_MEMPOOLLIST_LOCK ) {
-      // In the initial look at the head, the freelist wasn't locked.
-
-      // Proactively assign next pointer assuming a successful insertion into
-      // the list.
-      *reinterpret_cast<void*volatile*>(arg_alloc) = head ;
-
-      memory_fence();
-
-      // Attempt to insert at head of list.  If the list was changed
-      // (including being locked) between the initial look and now, head will
-      // be different than old_head.  This means the insert can't proceed and
-      // has to be tried again.
-
-      pending = head != atomic_compare_exchange( m_head_list, head, arg_alloc );
-    }
-  }
-}
-
-
-} // namespace Impl
-} // namespace Experimental
-} // namespace Kokkos
-
-#undef KOKKOS_MEMPOOLLIST_LOCK
-#undef KOKKOS_MEMPOOLLIST_INLINE
-
-#endif /* #ifndef KOKKOS_MEMORYPOOL_CPP */
-
+#include<impl/Kokkos_MemoryPool_Inline.hpp>
